changed latitude and longitude to double

+++ API/src/main/java/core/controllers/AccountController.java
@@ -68,7 +68,7 @@ public class AccountController {
     }
 
     @RequestMapping(value = "/newLot", method = RequestMethod.POST)
-    public Account newLot(String id, String name, String address, Float latitude, Float longitude, String price, int capacity) {
+    public Account newLot(String id, String name, String address, Double latitude, Double longitude, String price, int capacity) {
         Account account = accountRepo.findById(id);
         if (account == null) {
             System.out.println("Account with id " + id + " was not found.");
